+++ src/Service/Generator/IdStrategyInterface.php
@@ -3,6 +3,9 @@ declare(strict_types=1);
 
 namespace Shoman4eg\Nalog\Service\Generator;
 
+/**
+ * @author Artem Dubinin <artem@dubinin.me>
+ */
 interface IdStrategyInterface
 {
     public function getId(): string;