+++ CHANGELOG.md
@@ -4,7 +4,32 @@
 All notable changes to this project will be documented in this file.
 <!--- END HEADER -->
 
-## [0.1.0](https://github.com/shoman4eg/moy-nalog/compare/...v0.1.0) (2022-02-01)
+## [0.2.0](https://github.com/shoman4eg/moy-nalog/compare/v0.1.0...v0.2.0) (2022-04-09)
+### Features
+* Add Payment type methods
+* Add Cancel income method
+* Add custom client for create income
+* Add custom client for create income
+
+### Tests
+* Add tests
+* Add test for api client with custom access token
+* Add test for create income
+* Add test validation for income
+* Add test for cancel income
+* Add test for get paymentType
+
+### Documentation
+* Update Create api client
+* Update Create income
+* Update Get receipt
+* Update Get user info
+* Add Cancel income
+* Add References
+
+---
+
+## [0.1.0](https://github.com/shoman4eg/moy-nalog/compare/306901e41d3ae4d4a4913f6da9606213f9d9a11d...v0.1.0) (2022-02-01)
 
 ---
 Initial release